python/opendataloader-pdf/pyproject.toml
[project]
name = "opendataloader-pdf"
version = "0.0.0"
description = "A Python wrapper for the opendataloader-pdf Java CLI."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "opendataloader-project", email = "open.dataloader@hancom.com" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = []

[project.optional-dependencies]
hybrid = [
    "docling[easyocr]>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "python-multipart>=0.0.22",
]
gemini = [
    "google-genai>=0.3.0",
]
cloudrun = [
    "docling[easyocr]>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "python-multipart>=0.0.22",
    "google-genai>=0.3.0",
]

[project.scripts]
opendataloader-pdf = "opendataloader_pdf.wrapper:main"
opendataloader-pdf-hybrid = "opendataloader_pdf.hybrid_server:main"

[project.urls]
Homepage = "https://github.com/opendataloader-project/opendataloader-pdf"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/opendataloader_pdf"]
artifacts = [
    "src/opendataloader_pdf/jar/*.jar",
    "src/opendataloader_pdf/LICENSE",
    "src/opendataloader_pdf/NOTICE",
    "src/opendataloader_pdf/THIRD_PARTY/**",
]

[tool.hatch.build.targets.sdist]
include = [
    "src/opendataloader_pdf/**",
    "README.md",
    "hatch_build.py",
]

[tool.hatch.build.hooks.custom]
path = "hatch_build.py"

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[tool.uv]
package = true

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = ["dist", "build"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []
